Support SourceLocations in files larger than 4GB

+++ include/slang/diagnostics/DiagnosticClient.h
@@ -24,7 +24,7 @@ protected:
     const SourceManager* sourceManager = nullptr;
 
     void getIncludeStack(BufferID buffer, SmallVector<SourceLocation>& stack) const;
-    string_view getSourceLine(SourceLocation location, uint32_t col) const;
+    string_view getSourceLine(SourceLocation location, size_t col) const;
     static string_view getSeverityString(DiagnosticSeverity severity);
 };
